Day 4, part 2
Create another extension function, this time for overlaps.
The overlaps function is a bit half-assed because it's not
reflexive, but the surrounding stream already runs it both ways
so this doesn't matter.

The stream is basically the same, but it's so simple that it's
not worth extracting into a separate method and passing a
predicate (I tried, and it's easier just to duplicate the two
lines).