Day 8, part 2
A fairly straightforward solution for part 2.
The theoretical worst case is 10,000 (100x100) map accesses for each
tree in the 100x100 forest. However, that will only happen for a tree
of height 9 where all of the other trees in the row and column are
shorter — and by definition this can't happen very often. In
practice, the solution is fast.