Day 12, part 2
Brute force Djikstra's for each potential start point (there are 739
total in the input file).

There's a noticeable lag to calculate the answer, but it's not too
bad, probably just under a second. However, there's an NPE thrown
for each starting point that can't reach the end point, and that
sucks.