Part 2 is now lightning fast
Using a single invocation of Djikstra's and finding the distance to
every reachable square of height zero is basically instantaneous.

Some squares of height zero do not have a path to the end point, so
we need to filter that in our revised Djikstra's loop, because it
doesn't have an "if we've reached the endpoint, stop" check in it,
since we're looking for more than one possible end point.