Add extra test
Simplifying the two loops is going to be annoying (they
don't just have different predicates — one has an extra
break statement that the other doesn't have for points that
"fall to the abyss").

So, rather than unifying the two loops into a single
function, add a big old test that asserts that the state
for part 2's function matches the requirements.