Day 15, part 1 with an efficient solution
Start by attempting a straightforward approach:
generate all the points from each sensor that are known
not to contain beacons (because they are closer than the
closest known beacon), constrain to row Y, add them to a
set, and remove beacons actually on that line.

However, the first sensor in the prod data set is over a
million squares from its beacon, so this will take forever,
and we need something more efficient.

A given square, manhattan distance, and row will have either
2 (normal), 1 (the line is manhattan distance away from the
point), or 0 (the row is too far away) squares on the row.
Generate these as a sequence, add them to a set, remove
the points on the row that are actually beacons, and return
the count of the results.

This approach is almost instant.