Day 17, Part 2
Use the previous optimization of getting rid of irrelevant rocks
to build a cache key for cycle detection. The trick is to
"normalise" the fallenRocks set so that it doesn't depend on the
current y value.

Once a cycle is found, we can skip a bunch of work with some
simple multiplication and then do a few more iterations to get
to the desired number of rocks. The total solution finds a
solution for both parts in under 100ms.

Use a fork of Point with a 64-bit y value, and upgrade the
various height fields to be Longs instead of Ints.

Fix a nasty bug where the jet index and rock index were
calculating modulo on read, but incrementing the full value.
This a) breaks calculating the hash key for cycles impossible,
and b) prevents simulating more than 2^31 rocks correctly.