Performance tuning
Improve the performance of findInteriorPoints by
inlining isInteriorPoints and checking in bulk. This
allows us to maintain a list of known interior and known
exterior points that we've already checked, and as soon as
we reach one of those we know if the point is interior or
not.

With this improvement the answer is calculated in ~70ms.