Day 21, part 2
Instead of actually doing the equality operation, turn it into
a - b and look for a value that yields zero. This is basically
the same thing, but puts things in a more algebraic form.

And guess what — the big old function turns out to be LINEAR!!!
This means that we can predict it cheaply by finding f(x) for
two x values, determining the slope for y = mx + b, and then
solving for the y-intercept.

Well, almost. For whatever reason the function is *almost*
linear — the slope changes ever so slightly as x increases,
but the values for the production data are really huge and
it's big enough to make a difference. Maybe I should have been
using decimal division in part 1? Oh well.

Use a fairly huge value for x2 and a 128 decimal place
BigDecimal division to get a slope that predicts the result
accurately for the production and test data. Also, try +/-
a small amount and take the smallest value that returns 0 out
of the calculator function, just to be sure. Throw an
exception if this method doesn't work.

This method is nice and fast, calculating the answer in 60ms.